Use Transport Simulation Core from GitHub Packages

+++ common/src/main/java/org/mtr/block/BlockTrainAnnouncer.java
@@ -1,9 +1,5 @@
 package org.mtr.block;
 
-import it.unimi.dsi.fastutil.longs.Long2ObjectAVLTreeMap;
-import it.unimi.dsi.fastutil.longs.LongAVLTreeSet;
-import it.unimi.dsi.fastutil.longs.LongArrayList;
-import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import lombok.Getter;
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.BlockState;
@@ -17,6 +13,10 @@ import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
 import org.mtr.client.IDrawing;
 import org.mtr.core.tool.Utilities;
+import org.mtr.libraries.it.unimi.dsi.fastutil.longs.Long2ObjectAVLTreeMap;
+import org.mtr.libraries.it.unimi.dsi.fastutil.longs.LongAVLTreeSet;
+import org.mtr.libraries.it.unimi.dsi.fastutil.longs.LongArrayList;
+import org.mtr.libraries.it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import org.mtr.registry.BlockEntityTypes;
 
 import javax.annotation.Nonnull;
@@ -43,7 +43,7 @@ public class BlockTrainAnnouncer extends BlockTrainSensorBase {
 		QUEUE.forEach((time, tasks) -> {
 			if (time <= currentMillis) {
 				tasks.forEach(Runnable::run);
-				itemsToRemove.add(time.longValue());
+				itemsToRemove.add(time);
 			}
 		});
 		itemsToRemove.forEach(QUEUE::remove);
